update-coronovirus-infos.sh: Keep track of changes of the Coronavirus infos
It's hard to keep track of the changes of the information given on the
site of sozialministerium.at. This script should be put in a cronjob or
systemd.timer and will then automatically keep track of the changes.

Disclaimer: This is quick and dirty hack. Use at your own risk. :)